REFACTOR move runHeartbeat onto its own file
