+++ main.go
@@ -3,10 +3,8 @@ package main
 import (
 	"fmt"
 	"log"
-	"net/http"
 	"os"
 	"path/filepath"
-	"strings"
 	"time"
 
 	"github.com/tarm/serial"
@@ -54,9 +52,7 @@ func NewLogger() (*Logger, error) {
 }
 
 const (
-	heartbeatEndpoint = "https://nosnch.in/2b7bdbea9e"
-	heartbeatInterval = 5 * time.Minute
-	timeFormat        = "2006-01-02T15:04:05.999999"
+	timeFormat = "2006-01-02T15:04:05.999999"
 
 	serialPort = "/dev/ttyUSB0" // Change to the serial/COM port of the tower light
 	baudRate   = 9600
@@ -197,42 +193,6 @@ func (l *Light) Clear() error {
 func sendCommand(port *serial.Port, cmd byte) error {
 	_, err := port.Write([]byte{cmd})
 	return err
-}
-
-func sendHeartbeat() error {
-	payload := strings.NewReader("m=just checking in")
-	resp, err := http.Post(heartbeatEndpoint, "application/x-www-form-urlencoded", payload)
-	if err != nil {
-		return fmt.Errorf("failed to send heartbeat:: %s", err.Error())
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode < 200 || resp.StatusCode > 299 {
-		return fmt.Errorf("heartbeat failed with status code: %d", resp.StatusCode)
-	}
-
-	return nil
-}
-
-func runHeartbeat() {
-	fmt.Printf("In runHeartbeat\n")
-	ticker := time.NewTicker(heartbeatInterval)
-	defer ticker.Stop()
-
-	fmt.Printf("Starting loop\n")
-	for {
-		fmt.Printf("Sending heartbeat\n")
-		if err := sendHeartbeat(); err != nil {
-			fmt.Printf("Heartbeat error: %s\n", err.Error())
-			log.Printf("Heartbeat error:: %s", err.Error())
-		} else {
-			fmt.Printf("Heartbeat sent successfully.\n")
-			log.Printf("Heartbeat sent successfully..")
-		}
-		fmt.Printf("Finshed sending heartbeat\n")
-		<-ticker.C
-	}
-	fmt.Printf("Exiting runHeartbeat\n")
 }
 
 func main() {
